move utils to fdr_
